requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.2
sqlparse==0.4.3
tomli==2.0.1
